+++ topiary/__init__.py
@@ -171,7 +171,7 @@ from .amino_acids import (
     encode_amino_acids,
 )
 
-__version__ = "5.54.1"
+__version__ = "5.54.2"
 
 __all__ = [
     "normalize_python_types",